+++ src/app/modules/nb-table-sorter/assets/i18n/en.ts
@@ -5,6 +5,8 @@ export const locale = {
 		SEARCH: 'search',
 		ROWS_PER_PAGE: 'rows per page',
 		SHOWING_X_OF_Y_ROWS: 'Showing {{ amount }} of {{ total }} rows',
-		NO_RESULTS_FOUND: 'No results found'
+		NO_RESULTS_FOUND: 'No results found',
+		BATCH_ACTIONS: 'Batch actions',
+		APPLY: 'Apply'
 	}
 }