+++ src/app/modules/nb-table-sorter/assets/i18n/es.ts
@@ -5,6 +5,8 @@ export const locale = {
 		SEARCH: 'buscar',
 		ROWS_PER_PAGE: 'Filas por página',
 		SHOWING_X_OF_Y_ROWS: 'Mostrando {{ amount }} de {{ total }} resultados',
-		NO_RESULTS_FOUND: 'No se han encontrado resultados'
+		NO_RESULTS_FOUND: 'No se han encontrado resultados',
+		BATCH_ACTIONS: 'Acciones en lote',
+		APPLY: 'Aplicar'
 	}
 }